Fix target attack form: weapon selector, ranged-only options, action cost radio conflict (#166)
- Move MeleeWeaponInfo to GameMechanics.Combat so it can be shared by TargetingAttackerData
- Embed AvailableWeapons in TargetingAttackerData (set by CombatContent.BuildAttackerData) so
  the weapon list travels with the interaction instead of through a fragile 4-hop event chain
- TargetingAttackerPanel reads weapons from InitialData.AvailableWeapons in OnParametersSet,
  giving the target-attack form the same weapon selector as the anonymous-attack AttackMode
- Wrap Moving and Called Shot checkboxes in @if (RangedAttack) — melee attacks no longer show them
- Give each ActionCostSelector instance a unique radio group name (groupName = Guid) so attacker
  and defender radios don't interfere with each other
